# contracts/openapi/gateway-api.json
{
  "components": {
    "schemas": {
      "Alert": {
        "properties": {
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "revision": {
            "default": 1,
            "title": "Revision",
            "type": "integer"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "sar_review_status": {
            "title": "Sar Review Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "txn_id": {
            "title": "Txn Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "txn_id",
          "risk_score",
          "status",
          "alert_type",
          "created_at",
          "updated_at",
          "sar_review_status"
        ],
        "title": "Alert",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransactionDetails": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "counterparty_country": {
            "title": "Counterparty Country",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Features",
            "type": "object"
          },
          "risk_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Category"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "scorer_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scorer Version"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "txn_id": {
            "title": "Txn Id",
            "type": "string"
          }
        },
        "required": [
          "txn_id",
          "account_id",
          "amount",
          "currency",
          "timestamp",
          "counterparty_country",
          "features",
          "alerts"
        ],
        "title": "TransactionDetails",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Authenticated external gateway for the AML reference pipeline",
    "title": "AML Gateway API",
    "version": "3.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/transactions/{txn_id}": {
      "get": {
        "operationId": "get_transaction_v1_transactions__txn_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetails"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Transaction"
      }
    }
  }
}